--- asm/code_8095014.s
@@ -1,8 +1,0 @@
-	#include "asm/constants/gba_constants.inc"
-	#include "asm/macros.inc"
-
-	.syntax unified
-
-	.text
-
-        .align 2,0
