+++ asm/friend_area_action_menu.s
@@ -420,56 +420,4 @@ _0802778E:
 	bx r0
 	thumb_func_end sub_80276A8
 
-	thumb_func_start sub_8027794
-sub_8027794:
-	push {r4,lr}
-	ldr r4, _080277F0
-	ldr r0, [r4]
-	movs r1, 0xB6
-	lsls r1, 1
-	adds r0, r1
-	movs r1, 0
-	movs r2, 0x14
-	bl MemoryFill16
-	ldr r4, [r4]
-	movs r3, 0
-	movs r0, 0x8E
-	lsls r0, 1
-	adds r1, r4, r0
-	ldr r0, _080277F4
-	ldr r0, [r0]
-	str r0, [r1]
-	movs r0, 0x90
-	lsls r0, 1
-	adds r1, r4, r0
-	movs r0, 0x2
-	str r0, [r1]
-	movs r2, 0x1
-	movs r0, 0x92
-	lsls r0, 1
-	adds r1, r4, r0
-	ldr r0, _080277F8
-	ldr r0, [r0]
-	str r0, [r1]
-	movs r0, 0x94
-	lsls r0, 1
-	adds r1, r4, r0
-	movs r0, 0x3
-	str r0, [r1]
-	movs r1, 0x96
-	lsls r1, 1
-	adds r0, r4, r1
-	str r3, [r0]
-	adds r1, 0x4
-	adds r0, r4, r1
-	str r2, [r0]
-	pop {r4}
-	pop {r0}
-	bx r0
-	.align 2, 0
-_080277F0: .4byte gUnknown_203B2BC
-_080277F4: .4byte gUnknown_80D4920
-_080277F8: .4byte gUnknown_80D4928
-	thumb_func_end sub_8027794
-
         .align 2,0
