+++ ld_script.txt
@@ -181,8 +181,6 @@ SECTIONS {
         src/debug_menu.o(.text);
         src/debug_menu_mid.o(.text);
         asm/debug_menu_1.o(.text);
-        src/debug_menu_2.o(.text);
-        asm/debug_menu_2.o(.text);
         src/debug_menu_1.o(.text);
         asm/code_803B050.o(.text);
         src/code_803B050.o(.text);
